requirements.txt
flask>=3.0
flask-socketio>=5.3
pyodbc>=5.1
python-dotenv>=1.0
